python/requirements.txt
pandas==2.0.3
numpy==1.24.3
openpyxl==3.1.2
mysql-connector-python==8.2.0
xgboost==2.0.3
scikit-learn==1.3.2
joblib==1.3.2